Typo in Tool

+++ CS321-Project-Code/src/CS321/Project/Code/Tool.java
@@ -12,8 +12,8 @@ public class Tool extends Item {
   }
 
   @Override
-  public void takeDamage(int amt) {
-
+  public void takeDamage(double amt) {
+	  
   }
 
   @Override
